Update dibs spreadsheet key
Bootstrapping the project on my local machine, I realized the
spreadhseet that Nicole Cohen shared with me when we met was
different than the one whose ID appears in the settings variable.

We chatted on slack and she said she didn't have access to the
sheet referenced in the settings.  So, I updated the setting to
reflect the sheet where she has been working.