Document test books spreadsheet
When digging into multiple versions of a books spreadsheet, which
resulted in commit 50e61da7716083ff42bc8561a144a93b33ddfc87, I
realized that there was a final and test version of the spreadsheet
because the books team doesn't enter ISBNs into the final version
until later on in the process but the rig needs ISBNs to get the
book covers from Baker & Taylor and properly render.

Document this knowledge in the README along with a strategy for
overriding the setting that avoids toggling values in a file
that gets committed.