Start documenting project lifecycle
For annual projects like this one, I think it's really important
to document the set up process and other beats that happen when
the project is replicated each year.  Create a section in the
README for this and add the steps that I've encountered so far.
Add `TODO` comments for things that are missing.