Update tasks to reflect new station content workflow
Update the Fabric tasks dealing with external links to reflect the
actual station content workflow.

After seeing how Petra was editing the station content links, I needed
to change how the tasks work.  Update the tasks to reflect the fact that
the station content to be included is copied to and edited in a separate
spreadsheet.

Also update the generated link code to match the format from last year,
that is, wrapped in `<li>` tags and with the station name outside of the
link.