Support forcing use of NPR covers
For some reason the cover of "White Fur" came over as a pre-release
version. The NPR books version was correct, however.  I added a
setting to specify ISBNs of books for which the NPR version of the cover
image should always be used, regardless of whether a cover is available
from Baker and Taylor.